pyproject.toml
[tool.poetry]
name = "swe_4713_project"
version = "0.1.0"
description = "Initial attempt to add dependency management"
authors = ["Brandon Dunson <brandondunson15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
Flask-Login = "^0.5.0"
Flask-Mail = "^0.9.1"
Flask-SQLAlchemy = "^2.5.1"
werkzeug = "^2.0.1"
jwt = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
